# Replace generate_window with keras functionality