- Replaces generate_window with keras functionality
- Sets 10000 as default batch_size on predict. Batch_size on predict is *not* the same as the on for training, and is purely a performance trade-of (memory vs speed). Thus it is not customizable (at least not yet) from yaml. Did some rough benchmarks with different loookback-windows (144 / 72 / 36) and n_features (3,6) on transforming 100000 datapoints, and using a batch-size around 10000 was the fastest in general. 

- The previous implementation was equivalent with having batch_size=1 on predict, and that is very much slower than having it higher. Because of this we get this performance boost (measured on my laptop):
`model.transform(np.random.rand(1000,3))`
Old:
1min 14s  ± 1.7s per loop
new: 
132 ms ± 47.1 ms per loop
- Currently there is a small change in that in `fit` the first sample will be used twice. Can change that by removing it, but do we care? (see TODO)
- Made it such that all calls to `fit` calls `super().fit`, which generates a completely fresh model. This ensures that we are compatible with sklearn, especially for cross-validation, which of course expects a fresh model for each of the folds.
- There is also the change that we fit on all the data, and not only on the first 1/32th of the data, which we did before (bug)
- Removed the ability to output without a `y`, since it was not used.
- The new method might use a bit more memory. The reason is that we have to insert upto several rows at the start and end of the dataset for `TimeseriesGenerator` to generate the lookahead, and that probably causes numpy to copy the whole matrix. My current feel is that we should be OK with this, and rather implement something smarter if we need it later.

This closes #293